requirements.txt
# NEET Knowledge RAG Project
# A multi-format RAG system for NEET aspirants in India

# Core RAG Framework
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0

# Vector Database (FAISS)
faiss-cpu>=1.7.0

# Document Processing
pymupdf>=1.24.0  # PDF processing
pypdf>=4.0.0     # PDF reading
pytesseract>=0.3.10  # OCR for scanned PDFs
pdf2image>=1.17.0   # Convert PDF to images for OCR
pillow>=10.0.0   # Image processing

# YouTube & Video
youtube-transcript-api>=0.6.0  # YouTube transcripts
yt-dlp>=2024.0.0   # Video downloading
webvtt-py>=0.5.0    # VTT subtitle parsing

# Web Content
beautifulsoup4>=4.12.0  # HTML parsing
requests>=2.31.0   # HTTP requests

# Text Processing
python-dotenv>=1.0.0  # Environment variables
tiktoken>=0.7.0    # Token counting

# LLM Providers
openai>=1.0.0   # OpenAI / OpenRouter API

# Frontend
streamlit>=1.30.0

# Utilities
numpy>=1.24.0
pydantic>=2.0.0
pyyaml>=6.0
sentence-transformers>=2.2.0
